fix: Fix generation of "_generated" directory path
This was not being anchored to the project root, so when clearing the
cache in the terminal, the _generated directory was created in the
correct location, but when this was triggered from the front controller
it would place the src/_generated directory inside the public/ directory
which is not correct.

This commit simply defines the project root and adds this to the path as
as prefix.